+++ apps/openkey-example/CHANGELOG.md
@@ -1,5 +1,12 @@
 # tinycloud-openkey-example-app
 
+## 0.0.28-beta.5
+
+### Patch Changes
+
+- Updated dependencies [f0842d8]
+  - @tinycloud/web-sdk@2.11.0-beta.5
+
 ## 0.0.28-beta.4
 
 ### Patch Changes