+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tinycloud/cli
 
+## 0.8.3-beta.4
+
+### Patch Changes
+
+- Updated dependencies [f0842d8]
+  - @tinycloud/node-sdk@2.11.0-beta.5
+  - @tinycloud/operations@0.3.2-beta.4
+
 ## 0.8.3-beta.3
 
 ### Patch Changes