+++ packages/server/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @tinycloud/server
 
+## 2.4.11-beta.4
+
+### Patch Changes
+
+- Updated dependencies [f0842d8]
+  - @tinycloud/sdk-core@2.11.0-beta.5
+  - @tinycloud/node-sdk@2.11.0-beta.5
+
 ## 2.4.11-beta.3
 
 ### Patch Changes